+++ examples/hosts/app-integration/external-controller/src/app.ts
@@ -24,11 +24,11 @@ document.title = documentId;
 const dataObjectId = "dice";
 
 async function start(): Promise<void> {
-    const tinyliciousService = new TinyliciousService();
+    const service = new TinyliciousService();
     // Get or create the document
     const fluidDocument = createNew
-        ? await Fluid.createDocument(tinyliciousService, documentId, [KeyValueInstantiationFactory.registryEntry])
-        : await Fluid.getDocument(tinyliciousService, documentId, [KeyValueInstantiationFactory.registryEntry]);
+        ? await Fluid.createDocument(service, documentId, [KeyValueInstantiationFactory.registryEntry])
+        : await Fluid.getDocument(service, documentId, [KeyValueInstantiationFactory.registryEntry]);
 
     // We'll create the data object when we create the new document.
     const keyValueDataObject: IKeyValueDataObject = createNew
